requirements.txt
matplotlib==3.5.2
numpy==1.23.0
pandas==1.4.3
torch==1.11.0
optuna==2.10.1
seaborn==0.11.2
pytorch-lightning==1.6.4
PyYAML==6.0